Remove autopilot support from prepopulateForm, further rejiggery
