+++ plugin-hrm-form/src/HrmFormPlugin.tsx
@@ -45,7 +45,7 @@ import { subscribeNewMessageAlertOnPluginInit } from './notifications/newMessage
 import { subscribeReservedTaskAlert } from './notifications/reservedTask';
 import { setUpCounselorToolkits } from './components/toolkits/setUpCounselorToolkits';
 import { setupConferenceComponents, setUpConferenceActions } from './conference';
-import { setUpTransfers } from './transfer';
+import { setUpTransferActions } from './transfer/setUpTransferActions';
 
 const PLUGIN_NAME = 'HrmFormPlugin';
 
@@ -146,7 +146,6 @@ const setUpActions = (
   ActionFunctions.setUpPostSurvey(featureFlags);
 
   // bind setupObject to the functions that requires some initialization
-  const transferOverride = ActionFunctions.customTransferTask(setupObject);
   const wrapupOverride = ActionFunctions.wrapupTask(setupObject, getMessage);
   const beforeCompleteAction = ActionFunctions.beforeCompleteTask(featureFlags);
   // const afterWrapupAction = ActionFunctions.afterWrapupTask(featureFlags, setupObject);
@@ -155,10 +154,7 @@ const setUpActions = (
 
   Flex.Actions.addListener('afterAcceptTask', ActionFunctions.afterAcceptTask(featureFlags, setupObject, getMessage));
 
-  if (featureFlags.enable_transfers) Flex.Actions.replaceAction('TransferTask', transferOverride);
-
-  if (featureFlags.enable_transfers)
-    Flex.Actions.addListener('afterCancelTransfer', ActionFunctions.afterCancelTransfer);
+  setUpTransferActions(featureFlags.enable_transfers, setupObject);
 
   Flex.Actions.replaceAction('HangupCall', ActionFunctions.hangupCall);
 
@@ -205,7 +201,6 @@ export default class HrmFormPlugin extends FlexPlugin {
     ActionFunctions.loadCurrentDefinitionVersion();
 
     setUpSharedStateClient();
-    if (featureFlags.enable_transfers) setUpTransfers();
     setUpComponents(featureFlags, config, translateUI);
     setUpActions(featureFlags, config, getMessage);
