+++ plugin-hrm-form/src/utils/prepopulateForm.ts
@@ -18,7 +18,6 @@ import { ITask, Manager } from '@twilio/flex-ui';
 import { capitalize } from 'lodash';
 import { callTypes, FormDefinition, FormItemDefinition } from 'hrm-form-definitions';
 
-import { LexMemory, AutopilotMemory, FeatureFlags } from '../types/types';
 import { mapAge, mapGenericOption } from './mappers';
 import * as RoutingActions from '../states/routing/actions';
 import { prepopulateForm as prepopulateFormAction } from '../states/contacts/actions';
@@ -144,38 +143,8 @@ export const getValuesFromPreEngagementData = (
   return values;
 };
 
-type PreSurveyAnswers = LexMemory;
-type AutopilotAnswers = AutopilotMemory['twilio']['collected_data']['collect_survey']['answers'];
-
-const transformAnswers = (answers: AutopilotAnswers): PreSurveyAnswers => {
-  /**
-   * Map naming mismatches between autopilot and lex
-   * TODO: rename Lex 'slots' so we don't need to use this map
-   */
-  const keyRenames = {
-    // eslint-disable-next-line camelcase
-    about_self: 'aboutSelf',
-  };
-
-  const renameKey = key => (Object.keys(keyRenames).includes(key) ? keyRenames[key] : key);
-  return Object.keys(answers).reduce(
-    (acc, currentKey) => ({ ...acc, [renameKey(currentKey)]: answers[currentKey].answer }),
-    {},
-  );
-};
-
-const getAnswers = (isLexMemory: boolean, memory: LexMemory | AutopilotMemory): PreSurveyAnswers => {
-  if (isLexMemory) {
-    return memory as LexMemory;
-  }
-
-  // This can be removed after every helpline is using Lex
-  const autopilotAnswers = (memory as AutopilotMemory).twilio.collected_data.collect_survey.answers;
-  return transformAnswers(autopilotAnswers);
-};
-
 // eslint-disable-next-line sonarjs/cognitive-complexity
-export const prepopulateForm = (task: ITask, featureFlags: FeatureFlags) => {
+export const prepopulateForm = (task: ITask) => {
   const { memory, preEngagementData } = task.attributes;
 
   if (!memory && !preEngagementData) return;
@@ -208,17 +177,15 @@ export const prepopulateForm = (task: ITask, featureFlags: FeatureFlags) => {
     return;
   }
 
-  const answers = getAnswers(featureFlags.enable_lex, memory);
-
-  const isValidSurvey = Boolean(answers.aboutSelf); // determines if the memory has valid values or if it was aborted
-  const isAboutSelf = answers.aboutSelf === 'Yes';
+  const isValidSurvey = Boolean(memory.aboutSelf); // determines if the memory has valid values or if it was aborted
+  const isAboutSelf = memory.aboutSelf === 'Yes';
   // eslint-disable-next-line no-nested-ternary
   const callType = isValidSurvey ? (isAboutSelf ? callTypes.child : callTypes.caller) : null;
   const tabFormDefinition = isAboutSelf ? ChildInformationTab : CallerInformationTab;
   const prepopulateSurveyKeys = isAboutSelf ? survey.ChildInformationTab : survey.CallerInformationTab;
   const subroute = isAboutSelf ? 'childInformation' : 'callerInformation';
 
-  const surveyValues = getValuesFromAnswers(task, answers, tabFormDefinition, prepopulateSurveyKeys);
+  const surveyValues = getValuesFromAnswers(task, memory, tabFormDefinition, prepopulateSurveyKeys);
 
   // When a helpline has survey and no preEnagagement form
   if (memory && !preEngagementData) {
